feat: able to set gui-stream log level via JSON post
