+++ backend/gatterserver/api.py
@@ -5,14 +5,18 @@ from starlette.middleware.cors import CORSMiddleware
 
 from gatterserver.ble.discovery import BLEDiscoveryManager
 from gatterserver.emitters.emittermanager import EmitterManager
+from gatterserver.loggers import AsyncStreamHandler
 from gatterserver.routers import ble, jsonposts, tests, websockets
+
+# define and register global objects
 
 discovery_manager = BLEDiscoveryManager()
 emitter_manager = EmitterManager()
+gui_handler = AsyncStreamHandler()
 
 ble.register(discovery_manager)
-websockets.register(emitter_manager)
-jsonposts.register(emitter_manager)
+websockets.register(emitter_manager, gui_handler)
+jsonposts.register(emitter_manager, gui_handler)
 
 app = FastAPI()
